relocate various url.ast.data.path.* stuff to url.ast.data.predicate
