+++ ermrest/url/ast/data/makefile-vars
@@ -1,7 +1,8 @@
 
 ERMREST_URL_AST_DATA_PYTHON_FILES= \
 	__init__.py \
-	path.py
+	path.py \
+	predicate.py
 
 ERMREST_URL_AST_DATA_PYTHON_FILES_INSTALL=$(ERMREST_URL_AST_DATA_PYTHON_FILES:%=$(PYLIBDIR)/ermrest/url/ast/data/%)
